# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Payment Registration System API for Database Management Course UNLP",
        "title": "Payment Registration System",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "https://github.com/GabrielEValenzuela/Payment-Registration-System"
        },
        "license": {
            "name": "MIT License",
            "url": "http://www.apache.org/licenses/MIT.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/cards/expiring/{day}/{month}/{year}": {
            "get": {
                "description": "Retrieves the cards that will expire in the next 30 days based on the current day, month, and year.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Get cards expiring in the next 30 days",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Day",
                        "name": "day",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cards expiring in the next 30 days",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve expiring cards",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/cards/purchase/monthly": {
            "get": {
                "description": "Retrieves the monthly purchase details for the specified card based on CUIT and payment voucher.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Get monthly purchase details for a card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CUIT",
                        "name": "cuit",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Final amount",
                        "name": "finalAmount",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment voucher",
                        "name": "paymentVoucher",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly purchase details retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve monthly purchase details",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/cards/summary/{cardNumber}/{month}/{year}": {
            "get": {
                "description": "Retrieves the payment summary for the specified card number, month, and year.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Get the payment summary for a card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card number",
                        "name": "cardNumber",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment summary retrieved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve payment summary",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/cards/top": {
            "get": {
                "description": "Retrieves the top 10 cards by purchases.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Get top 10 cards by purchases",
                "responses": {
                    "200": {
                        "description": "Top 10 cards by purchases",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve top 10 cards",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    }
}